refactor datetimeparser
