+++ src/main/java/seedu/duke/data/Storage.java
@@ -18,7 +18,7 @@ import java.io.BufferedWriter;
 import java.io.IOException;
 import java.io.FileWriter;
 
-import static seedu.duke.command.DateTimeParser.getDateTime;
+import static seedu.duke.parser.DateTimeParser.getDateTime;
 
 /**
  * A class that stores current task list and loads it on request from disc.
