+++ src/main/java/seedu/duke/task/Deadline.java
@@ -3,8 +3,8 @@ package seedu.duke.task;
 import seedu.duke.ui.Ui;
 import java.time.LocalDateTime;
 
-import static seedu.duke.command.DateTimeParser.convertDateTime;
-import static seedu.duke.command.DateTimeParser.toDateTimeString;
+import static seedu.duke.parser.DateTimeParser.convertDateTime;
+import static seedu.duke.parser.DateTimeParser.toDateTimeString;
 
 /**
  * This task type inherits from Task.
